% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressnet.R
\name{encode_text}
\alias{encode_text}
\title{Encodes text for input into the neural network}
\usage{
encode_text(txt)
}
\arguments{
\item{txt}{a vector of address strings}
}
\value{
a list containing the original string lengths and a padded matrix encoding the addresses
}
\description{
Encodes text for input into the neural network
}
