% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressnet.R
\name{parse_addresses}
\alias{parse_addresses}
\title{Converts a vector of addresses into a table showing each address component}
\usage{
parse_addresses(address)
}
\arguments{
\item{address}{a vector of addresses}
}
\value{
a tibble of address components
}
\description{
Converts a vector of addresses into a table showing each address component
}
